requirements.txt
paramiko
termcolor
python-dotenv
scp
r2pipe
nltk
pexpect
requests